initBase/requirements.txt
gunicorn
django==4.2.16
psycopg2-binary


#api setup
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support


celery[redis]

django_celery_results

requests
icecream

# Auth
PyJWT
